requirements.txt
alpaca-py==0.40.1
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
idna==3.10
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
msgpack==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.3.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sseclient-py==1.8.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
fastapi
uvicorn[standard]
psutil
python-telegram-bot